Implement `createApp` API and define related types
The `createApp` function was added to enable the creation of application instances with a render method. This procedure replaced the `helloChibivue` function in the index file. Along with this, "Options" and "App" types were implemented to support this functionality. The related changes have been incorporated in the playground's `main.ts` file.