Add reactivity functionality to codebase
The reactivity package has been added to enhance the functionality of the application. This module includes the creation of a new 'Dep' type and corresponding functions, as well as a reactivity setup enabling change tracking and triggering on objects. This update promotes better data flow and update handling within the application.