Refactor VNode interface to include generic HostNode
The VNode interface has been enhanced to include a generic HostNode parameter in runtime-core package. Along with the host node, this modification entails updates to '' type that can now accept null values, and 'children' type is defined as 'VNodeNormalizedChildren'. The 'normalizeVNode' function is also being utilized to convert 'VNodeChild' into the updated 'VNode'.