Update VNode type and component inclusion in runtime-core
This commit broadens the `type` parameter within the `h` function in `h.ts` and the `VNodeTypes` array in `vnode.ts` to include not only strings but also objects. Moreover, it adds `component: ComponentInternalInstance | null` in `VNode` interface. These changes aim to increase code flexibility and enhance handling of VNode and Component instances.