Add runtime compiler registration and compile function
Introduced a runtime compiler registration and a compile function to facilitate template conversion to a render function. Integrated the registration of these changes to the main index file and added exports to the compiler-dom package for greater functionality. A compile function has also been defined to facilitate the template-to-render function conversion.