Add event listener to button in main.ts
The code has been modified to add an event listener to a button within the main.ts file. When this button is clicked, an exclamation mark is appended to the text content of the first heading. A button with the label 'Click me!' has also been introduced to the template.