Adjust setup method in ComponentInstance
The adjustment of the setup method in the ComponentInstance now enables it to handle a result being a function or an object. If the setup method's return value is a function, it gets assigned to the instance's render method. If it's an object and not null, it gets assigned to the newly added setupState property.