Refactor setup method in ComponentInternalInstance
Restructured the setup method in ComponentInternalInstance to maintain consistency and readability. Now, if the setup method's return value is a function, it's directly assigned to the instance's render method. If the response is an object (and not null), it's assigned to the newly added setupState property. Also, rendering logic for both function and object return types has been improved.