Implement compiler-dom in vite-plugin-chibivue
The compiler-dom method is imported and used to compile the template content in vite-plugin-chibivue. This provides a way to handle .vue files and compile them according to the new CompilerOptions object, which assumes a browser environment as default. Consequently, the transform function now returns the new compiled code instead of an empty object.