Update the "vite" package version and remove dev dependencies
In an effort to streamline the project's dependencies, the version of the "vite" package has been updated. All dependencies previously marked as development dependencies are now marked as regular dependencies, leading to modifications in the pnpm-lock file to reflect these changes. Changes in the package.json file reflect the movement of "vite" from devDependencies to dependencies.